requirements.txt
openenv-core>=0.1.0
pydantic>=2.0.0
openai>=1.0.0
pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
